requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pillow==10.4.0
requests==2.32.3
markdown==3.7
bleach==6.1.0
gunicorn==23.0.0